+++ bin/wits_flutter_cli.dart
@@ -1,5 +1,6 @@
 import 'package:wits_flutter_cli/wits_flutter_cli.dart';
 
 void main(List<String> arguments) {
-  runModule(arguments);
+  // runModule(arguments);
+  createConfig(arguments);
 }