+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 1.0.2
 
 - Minor Changes.
+
+## 1.0.3
+
+- Added function to create launch.json file for selected module
+
+## 1.0.4
+
+- Added error notifiers in case of missing arguments.